hyperactor_mesh::actor_mesh

Trait ActorMesh

pub trait ActorMesh: Mesh<Id = ActorMeshId> {
    type Actor: Referable;

    // Required methods
    fn proc_mesh(&self) -> &ProcMesh;
    fn name(&self) -> &str;
    fn v1(&self) -> Option<ActorMeshRef<Self::Actor>>;

    // Provided methods
    fn cast<M>(
        &self,
        cx: &impl Actor,
        selection: Selection,
        message: M,
    ) -> Result<(), CastError>
       where Self::Actor: RemoteHandles<M> + RemoteHandles<IndexedErasedUnbound<M>>,
             M: Castable + RemoteMessage + Clone { ... }
    fn world_id(&self) -> &WorldId { ... }
    fn iter_actor_refs(&self) -> Box<dyn Iterator<Item = ActorRef<Self::Actor>>> { ... }
    fn stop<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        cx: &'life1 (impl 'async_trait + Actor),
        reason: &'life2 str,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn bind(&self) -> ActorMeshRef<Self::Actor> { ... }
}

A mesh of actors, all of which reside on the same ProcMesh.

Required Associated Types

type Actor: Referable

The type of actor in the mesh.

Required Methods

fn proc_mesh(&self) -> &ProcMesh

The ProcMesh on top of which this actor mesh is spawned.

fn name(&self) -> &str

The name given to the actors in this mesh.

fn v1(&self) -> Option<ActorMeshRef<Self::Actor>>

Retrieves the v1 mesh for this v0 ActorMesh, if it is available.

Provided Methods

fn cast<M>( &self, cx: &impl Actor, selection: Selection, message: M, ) -> Result<(), CastError>
where Self::Actor: RemoteHandles<M> + RemoteHandles<IndexedErasedUnbound<M>>, M: Castable + RemoteMessage + Clone,

Cast an M-typed message to the ranks selected by sel in this ActorMesh.

fn world_id(&self) -> &WorldId

fn iter_actor_refs(&self) -> Box<dyn Iterator<Item = ActorRef<Self::Actor>>>

Iterate over all ActorRef<Self::Actor> in this mesh.

fn stop<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, cx: &'life1 (impl 'async_trait + Actor), reason: &'life2 str, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

fn bind(&self) -> ActorMeshRef<Self::Actor>

Get a serializeable reference to this mesh similar to ActorHandle::bind

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.