hyperactor_mesh::alloc

Module remoteprocess

Modules

__mock_MockRemoteProcessAllocInitializer

__mock_MockRemoteProcessAllocInitializer_RemoteProcessAllocInitializer

Structs

MockRemoteProcessAllocInitializer

Interface to provide the set of hosts to be used by RemoteProcessAlloc.

RemoteProcessAlloc

A generalized implementation of an Alloc using one or more hosts running RemoteProcessAlloc for process allocation.

RemoteProcessAllocHost

A host entry passed down from the initializer.

RemoteProcessAllocator

Allocator with a service frontend that wraps ProcessAllocator.

Enums

RemoteProcessAllocatorMessage

Control messages sent from remote process allocator to local allocator.

RemoteProcessProcStateMessage

Control message sent from local allocator to remote allocator relaying process state updates. AsRefStr allows us to log the values

Traits

RemoteProcessAllocInitializer

Interface to provide the set of hosts to be used by RemoteProcessAlloc.