hyperactor_mesh::alloc

Trait Alloc

pub trait Alloc {
    // Required methods
    fn next<'life0, 'async_trait>(
        &'life0 mut self,
    ) -> Pin<Box<dyn Future<Output = Option<ProcState>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn spec(&self) -> &AllocSpec;
    fn extent(&self) -> &Extent;
    fn alloc_name(&self) -> &AllocName;
    fn stop<'life0, 'async_trait>(
        &'life0 mut self,
    ) -> Pin<Box<dyn Future<Output = Result<(), AllocatorError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided methods
    fn shape(&self) -> Shape { ... }
    fn transport(&self) -> ChannelTransport { ... }
    fn stop_and_wait<'life0, 'async_trait>(
        &'life0 mut self,
    ) -> Pin<Box<dyn Future<Output = Result<(), AllocatorError>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
    fn is_local(&self) -> bool { ... }
    fn client_router_addr(&self) -> ChannelAddr { ... }
}

An alloc is a specific allocation, returned by an Allocator.

Required Methods

fn next<'life0, 'async_trait>( &'life0 mut self, ) -> Pin<Box<dyn Future<Output = Option<ProcState>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the next proc event. None indicates that there are no more events, and that the alloc is stopped.

fn spec(&self) -> &AllocSpec

The spec against which this alloc is executing.

fn extent(&self) -> &Extent

The shape of the alloc.

fn alloc_name(&self) -> &AllocName

The name of this alloc, uniquely identifying it.

fn stop<'life0, 'async_trait>( &'life0 mut self, ) -> Pin<Box<dyn Future<Output = Result<(), AllocatorError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Stop this alloc, shutting down all of its procs. A clean shutdown should result in Stop events from all allocs, followed by the end of the event stream.

Provided Methods

fn shape(&self) -> Shape

The shape of the alloc. (Deprecated.)

fn transport(&self) -> ChannelTransport

The channel transport used the procs in this alloc.

fn stop_and_wait<'life0, 'async_trait>( &'life0 mut self, ) -> Pin<Box<dyn Future<Output = Result<(), AllocatorError>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Stop this alloc and wait for all procs to stop. Call will block until all ProcState events have been drained.

fn is_local(&self) -> bool

Returns whether the alloc is a local alloc: that is, its procs are not independent processes, but just threads in the selfsame process.

fn client_router_addr(&self) -> ChannelAddr

The address that should be used to serve the client’s router.