hyperactor_mesh::comm::multicast

Trait CastInfo

pub trait CastInfo {
    // Required methods
    fn cast_point(&self) -> Point;
    fn sender(&self) -> &ActorId;
}

Required Methods

fn cast_point(&self) -> Point

Get the cast rank and cast shape. If something wasn’t explicitly sent via a cast, then we represent it as the only member of a 0-dimensonal cast shape, which is the same as a singleton.

fn sender(&self) -> &ActorId