hyperactor_mesh::connect

Function accept

pub async fn accept<C: Actor + Clone>(
    caps: C,
    self_id: ActorAddr,
    message: Connect,
) -> Result<ActorConnection<C>>

Helper used by Handler<Connect>s to accept a connection initiated by a Connect message and return AsyncRead and AsyncWrite streams that can be used to communicate with the other side.