hyperactor_mesh

Enum Error

pub enum Error {

    InvalidRankCardinality {
        expected: usize,
        actual: usize,
    },
    NameParseError(NameParseError),
    HostMeshRefParseError(HostMeshRefParseError),
    AllocatorError(Box<AllocatorError>),
    ChannelError(Box<ChannelError>),
    MailboxError(Box<MailboxError>),
    CodecError(CodecError),
    ConfigurationError(Error),
    UnroutableMesh(),
    CallError(ActorId, Error),
    ActorTypeNotRegistered(String),
    GspawnError(Name, String),
    SendingError(ActorId, Box<MailboxSenderError>),
    CastingError(Name, Error),
    HostMeshAgentConfigurationError(ActorId, String),
    ProcCreationError {
        state: Box<State<ProcState>>,
        host_rank: usize,
        mesh_agent: ActorRef<HostAgent>,
    },
    ProcSpawnError {
        statuses: RankedValues<Status>,
    },
    ActorSpawnError {
        statuses: RankedValues<Status>,
    },
    ActorStopError {
        statuses: RankedValues<Status>,
    },
    SingletonActorSpawnError(Error),
    ControllerActorSpawnError(Name, Error),
    RankedProc(ProcId),
    Supervision(Box<MeshFailure>),
    NotExist(Name),
    Io(Error),
    Host(HostError),
    Other(Error),

}

Errors that occur during mesh operations.

Variants

InvalidRankCardinality

Fields

expected: usize

actual: usize

NameParseError(NameParseError)

HostMeshRefParseError(HostMeshRefParseError)

AllocatorError(Box<AllocatorError>)

ChannelError(Box<ChannelError>)

MailboxError(Box<MailboxError>)

CodecError(CodecError)

ConfigurationError(Error)

UnroutableMesh()

CallError(ActorId, Error)

ActorTypeNotRegistered(String)

GspawnError(Name, String)

SendingError(ActorId, Box<MailboxSenderError>)

CastingError(Name, Error)

HostMeshAgentConfigurationError(ActorId, String)

ProcCreationError

Fields

state: Box<State<ProcState>>

host_rank: usize

mesh_agent: ActorRef<HostAgent>

ProcSpawnError

Fields

statuses: RankedValues<Status>

ActorSpawnError

Fields

statuses: RankedValues<Status>

ActorStopError

Fields

statuses: RankedValues<Status>

SingletonActorSpawnError(Error)

ControllerActorSpawnError(Name, Error)

RankedProc(ProcId)

Supervision(Box<MeshFailure>)

NotExist(Name)

Io(Error)

Host(HostError)

Other(Error)

Implementations

impl Error

pub fn is_invalid_rank_cardinality(&self) -> bool

Returns true if this is a Error::InvalidRankCardinality, otherwise false

pub fn as_invalid_rank_cardinality_mut( &mut self, ) -> Option<(&mut usize, &mut usize)>

Optionally returns mutable references to the inner fields if this is a Error::InvalidRankCardinality, otherwise None

pub fn as_invalid_rank_cardinality(&self) -> Option<(&usize, &usize)>

Optionally returns references to the inner fields if this is a Error::InvalidRankCardinality, otherwise None

pub fn into_invalid_rank_cardinality(self) -> Result<(usize, usize), Self>

Returns the inner fields if this is a Error::InvalidRankCardinality, otherwise returns back the enum in the Err case of the result

pub fn is_name_parse_error(&self) -> bool

Returns true if this is a Error::NameParseError, otherwise false

pub fn as_name_parse_error_mut(&mut self) -> Option<&mut NameParseError>

Optionally returns mutable references to the inner fields if this is a Error::NameParseError, otherwise None

pub fn as_name_parse_error(&self) -> Option<&NameParseError>

Optionally returns references to the inner fields if this is a Error::NameParseError, otherwise None

pub fn into_name_parse_error(self) -> Result<NameParseError, Self>

Returns the inner fields if this is a Error::NameParseError, otherwise returns back the enum in the Err case of the result

pub fn is_host_mesh_ref_parse_error(&self) -> bool

Returns true if this is a Error::HostMeshRefParseError, otherwise false

pub fn as_host_mesh_ref_parse_error_mut( &mut self, ) -> Option<&mut HostMeshRefParseError>

Optionally returns mutable references to the inner fields if this is a Error::HostMeshRefParseError, otherwise None

pub fn as_host_mesh_ref_parse_error(&self) -> Option<&HostMeshRefParseError>

Optionally returns references to the inner fields if this is a Error::HostMeshRefParseError, otherwise None

pub fn into_host_mesh_ref_parse_error( self, ) -> Result<HostMeshRefParseError, Self>

Returns the inner fields if this is a Error::HostMeshRefParseError, otherwise returns back the enum in the Err case of the result

pub fn is_allocator_error(&self) -> bool

Returns true if this is a Error::AllocatorError, otherwise false

pub fn as_allocator_error_mut(&mut self) -> Option<&mut Box<AllocatorError>>

Optionally returns mutable references to the inner fields if this is a Error::AllocatorError, otherwise None

pub fn as_allocator_error(&self) -> Option<&Box<AllocatorError>>

Optionally returns references to the inner fields if this is a Error::AllocatorError, otherwise None

pub fn into_allocator_error(self) -> Result<Box<AllocatorError>, Self>

Returns the inner fields if this is a Error::AllocatorError, otherwise returns back the enum in the Err case of the result

pub fn is_channel_error(&self) -> bool

Returns true if this is a Error::ChannelError, otherwise false

pub fn as_channel_error_mut(&mut self) -> Option<&mut Box<ChannelError>>

Optionally returns mutable references to the inner fields if this is a Error::ChannelError, otherwise None

pub fn as_channel_error(&self) -> Option<&Box<ChannelError>>

Optionally returns references to the inner fields if this is a Error::ChannelError, otherwise None

pub fn into_channel_error(self) -> Result<Box<ChannelError>, Self>

Returns the inner fields if this is a Error::ChannelError, otherwise returns back the enum in the Err case of the result

pub fn is_mailbox_error(&self) -> bool

Returns true if this is a Error::MailboxError, otherwise false

pub fn as_mailbox_error_mut(&mut self) -> Option<&mut Box<MailboxError>>

Optionally returns mutable references to the inner fields if this is a Error::MailboxError, otherwise None

pub fn as_mailbox_error(&self) -> Option<&Box<MailboxError>>

Optionally returns references to the inner fields if this is a Error::MailboxError, otherwise None

pub fn into_mailbox_error(self) -> Result<Box<MailboxError>, Self>

Returns the inner fields if this is a Error::MailboxError, otherwise returns back the enum in the Err case of the result

pub fn is_codec_error(&self) -> bool

Returns true if this is a Error::CodecError, otherwise false

pub fn as_codec_error_mut(&mut self) -> Option<&mut CodecError>

Optionally returns mutable references to the inner fields if this is a Error::CodecError, otherwise None

pub fn as_codec_error(&self) -> Option<&CodecError>

Optionally returns references to the inner fields if this is a Error::CodecError, otherwise None

pub fn into_codec_error(self) -> Result<CodecError, Self>

Returns the inner fields if this is a Error::CodecError, otherwise returns back the enum in the Err case of the result

pub fn is_configuration_error(&self) -> bool

Returns true if this is a Error::ConfigurationError, otherwise false

pub fn as_configuration_error_mut(&mut self) -> Option<&mut Error>

Optionally returns mutable references to the inner fields if this is a Error::ConfigurationError, otherwise None

pub fn as_configuration_error(&self) -> Option<&Error>

Optionally returns references to the inner fields if this is a Error::ConfigurationError, otherwise None

pub fn into_configuration_error(self) -> Result<Error, Self>

Returns the inner fields if this is a Error::ConfigurationError, otherwise returns back the enum in the Err case of the result

pub fn is_unroutable_mesh(&self) -> bool

Returns true if this is a Error::UnroutableMesh, otherwise false

pub fn as_unroutable_mesh_mut(&mut self) -> Option<()>

Optionally returns mutable references to the inner fields if this is a Error::UnroutableMesh, otherwise None

pub fn as_unroutable_mesh(&self) -> Option<()>

Optionally returns references to the inner fields if this is a Error::UnroutableMesh, otherwise None

pub fn into_unroutable_mesh(self) -> Result<(), Self>

Returns the inner fields if this is a Error::UnroutableMesh, otherwise returns back the enum in the Err case of the result

pub fn is_call_error(&self) -> bool

Returns true if this is a Error::CallError, otherwise false

pub fn as_call_error_mut(&mut self) -> Option<(&mut ActorId, &mut Error)>

Optionally returns mutable references to the inner fields if this is a Error::CallError, otherwise None

pub fn as_call_error(&self) -> Option<(&ActorId, &Error)>

Optionally returns references to the inner fields if this is a Error::CallError, otherwise None

pub fn into_call_error(self) -> Result<(ActorId, Error), Self>

Returns the inner fields if this is a Error::CallError, otherwise returns back the enum in the Err case of the result

pub fn is_actor_type_not_registered(&self) -> bool

Returns true if this is a Error::ActorTypeNotRegistered, otherwise false

pub fn as_actor_type_not_registered_mut(&mut self) -> Option<&mut String>

Optionally returns mutable references to the inner fields if this is a Error::ActorTypeNotRegistered, otherwise None

pub fn as_actor_type_not_registered(&self) -> Option<&String>

Optionally returns references to the inner fields if this is a Error::ActorTypeNotRegistered, otherwise None

pub fn into_actor_type_not_registered(self) -> Result<String, Self>

Returns the inner fields if this is a Error::ActorTypeNotRegistered, otherwise returns back the enum in the Err case of the result

pub fn is_gspawn_error(&self) -> bool

Returns true if this is a Error::GspawnError, otherwise false

pub fn as_gspawn_error_mut(&mut self) -> Option<(&mut Name, &mut String)>

Optionally returns mutable references to the inner fields if this is a Error::GspawnError, otherwise None

pub fn as_gspawn_error(&self) -> Option<(&Name, &String)>

Optionally returns references to the inner fields if this is a Error::GspawnError, otherwise None

pub fn into_gspawn_error(self) -> Result<(Name, String), Self>

Returns the inner fields if this is a Error::GspawnError, otherwise returns back the enum in the Err case of the result

pub fn is_sending_error(&self) -> bool

Returns true if this is a Error::SendingError, otherwise false

pub fn as_sending_error_mut( &mut self, ) -> Option<(&mut ActorId, &mut Box<MailboxSenderError>)>

Optionally returns mutable references to the inner fields if this is a Error::SendingError, otherwise None

pub fn as_sending_error(&self) -> Option<(&ActorId, &Box<MailboxSenderError>)>

Optionally returns references to the inner fields if this is a Error::SendingError, otherwise None

pub fn into_sending_error( self, ) -> Result<(ActorId, Box<MailboxSenderError>), Self>

Returns the inner fields if this is a Error::SendingError, otherwise returns back the enum in the Err case of the result

pub fn is_casting_error(&self) -> bool

Returns true if this is a Error::CastingError, otherwise false

pub fn as_casting_error_mut(&mut self) -> Option<(&mut Name, &mut Error)>

Optionally returns mutable references to the inner fields if this is a Error::CastingError, otherwise None

pub fn as_casting_error(&self) -> Option<(&Name, &Error)>

Optionally returns references to the inner fields if this is a Error::CastingError, otherwise None

pub fn into_casting_error(self) -> Result<(Name, Error), Self>

Returns the inner fields if this is a Error::CastingError, otherwise returns back the enum in the Err case of the result

pub fn is_host_mesh_agent_configuration_error(&self) -> bool

Returns true if this is a Error::HostMeshAgentConfigurationError, otherwise false

pub fn as_host_mesh_agent_configuration_error_mut( &mut self, ) -> Option<(&mut ActorId, &mut String)>

Optionally returns mutable references to the inner fields if this is a Error::HostMeshAgentConfigurationError, otherwise None

pub fn as_host_mesh_agent_configuration_error( &self, ) -> Option<(&ActorId, &String)>

Optionally returns references to the inner fields if this is a Error::HostMeshAgentConfigurationError, otherwise None

pub fn into_host_mesh_agent_configuration_error( self, ) -> Result<(ActorId, String), Self>

Returns the inner fields if this is a Error::HostMeshAgentConfigurationError, otherwise returns back the enum in the Err case of the result

pub fn is_proc_creation_error(&self) -> bool

Returns true if this is a Error::ProcCreationError, otherwise false

pub fn as_proc_creation_error_mut( &mut self, ) -> Option<(&mut Box<State<ProcState>>, &mut usize, &mut ActorRef<HostAgent>)>

Optionally returns mutable references to the inner fields if this is a Error::ProcCreationError, otherwise None

pub fn as_proc_creation_error( &self, ) -> Option<(&Box<State<ProcState>>, &usize, &ActorRef<HostAgent>)>

Optionally returns references to the inner fields if this is a Error::ProcCreationError, otherwise None

pub fn into_proc_creation_error( self, ) -> Result<(Box<State<ProcState>>, usize, ActorRef<HostAgent>), Self>

Returns the inner fields if this is a Error::ProcCreationError, otherwise returns back the enum in the Err case of the result

pub fn is_proc_spawn_error(&self) -> bool

Returns true if this is a Error::ProcSpawnError, otherwise false

pub fn as_proc_spawn_error_mut(&mut self) -> Option<&mut RankedValues<Status>>

Optionally returns mutable references to the inner fields if this is a Error::ProcSpawnError, otherwise None

pub fn as_proc_spawn_error(&self) -> Option<&RankedValues<Status>>

Optionally returns references to the inner fields if this is a Error::ProcSpawnError, otherwise None

pub fn into_proc_spawn_error(self) -> Result<RankedValues<Status>, Self>

Returns the inner fields if this is a Error::ProcSpawnError, otherwise returns back the enum in the Err case of the result

pub fn is_actor_spawn_error(&self) -> bool

Returns true if this is a Error::ActorSpawnError, otherwise false

pub fn as_actor_spawn_error_mut(&mut self) -> Option<&mut RankedValues<Status>>

Optionally returns mutable references to the inner fields if this is a Error::ActorSpawnError, otherwise None

pub fn as_actor_spawn_error(&self) -> Option<&RankedValues<Status>>

Optionally returns references to the inner fields if this is a Error::ActorSpawnError, otherwise None

pub fn into_actor_spawn_error(self) -> Result<RankedValues<Status>, Self>

Returns the inner fields if this is a Error::ActorSpawnError, otherwise returns back the enum in the Err case of the result

pub fn is_actor_stop_error(&self) -> bool

Returns true if this is a Error::ActorStopError, otherwise false

pub fn as_actor_stop_error_mut(&mut self) -> Option<&mut RankedValues<Status>>

Optionally returns mutable references to the inner fields if this is a Error::ActorStopError, otherwise None

pub fn as_actor_stop_error(&self) -> Option<&RankedValues<Status>>

Optionally returns references to the inner fields if this is a Error::ActorStopError, otherwise None

pub fn into_actor_stop_error(self) -> Result<RankedValues<Status>, Self>

Returns the inner fields if this is a Error::ActorStopError, otherwise returns back the enum in the Err case of the result

pub fn is_singleton_actor_spawn_error(&self) -> bool

Returns true if this is a Error::SingletonActorSpawnError, otherwise false

pub fn as_singleton_actor_spawn_error_mut(&mut self) -> Option<&mut Error>

Optionally returns mutable references to the inner fields if this is a Error::SingletonActorSpawnError, otherwise None

pub fn as_singleton_actor_spawn_error(&self) -> Option<&Error>

Optionally returns references to the inner fields if this is a Error::SingletonActorSpawnError, otherwise None

pub fn into_singleton_actor_spawn_error(self) -> Result<Error, Self>

Returns the inner fields if this is a Error::SingletonActorSpawnError, otherwise returns back the enum in the Err case of the result

pub fn is_controller_actor_spawn_error(&self) -> bool

Returns true if this is a Error::ControllerActorSpawnError, otherwise false

pub fn as_controller_actor_spawn_error_mut( &mut self, ) -> Option<(&mut Name, &mut Error)>

Optionally returns mutable references to the inner fields if this is a Error::ControllerActorSpawnError, otherwise None

pub fn as_controller_actor_spawn_error(&self) -> Option<(&Name, &Error)>

Optionally returns references to the inner fields if this is a Error::ControllerActorSpawnError, otherwise None

pub fn into_controller_actor_spawn_error(self) -> Result<(Name, Error), Self>

Returns the inner fields if this is a Error::ControllerActorSpawnError, otherwise returns back the enum in the Err case of the result

pub fn is_ranked_proc(&self) -> bool

Returns true if this is a Error::RankedProc, otherwise false

pub fn as_ranked_proc_mut(&mut self) -> Option<&mut ProcId>

Optionally returns mutable references to the inner fields if this is a Error::RankedProc, otherwise None

pub fn as_ranked_proc(&self) -> Option<&ProcId>

Optionally returns references to the inner fields if this is a Error::RankedProc, otherwise None

pub fn into_ranked_proc(self) -> Result<ProcId, Self>

Returns the inner fields if this is a Error::RankedProc, otherwise returns back the enum in the Err case of the result

pub fn is_supervision(&self) -> bool

Returns true if this is a Error::Supervision, otherwise false

pub fn as_supervision_mut(&mut self) -> Option<&mut Box<MeshFailure>>

Optionally returns mutable references to the inner fields if this is a Error::Supervision, otherwise None

pub fn as_supervision(&self) -> Option<&Box<MeshFailure>>

Optionally returns references to the inner fields if this is a Error::Supervision, otherwise None

pub fn into_supervision(self) -> Result<Box<MeshFailure>, Self>

Returns the inner fields if this is a Error::Supervision, otherwise returns back the enum in the Err case of the result

pub fn is_not_exist(&self) -> bool

Returns true if this is a Error::NotExist, otherwise false

pub fn as_not_exist_mut(&mut self) -> Option<&mut Name>

Optionally returns mutable references to the inner fields if this is a Error::NotExist, otherwise None

pub fn as_not_exist(&self) -> Option<&Name>

Optionally returns references to the inner fields if this is a Error::NotExist, otherwise None

pub fn into_not_exist(self) -> Result<Name, Self>

Returns the inner fields if this is a Error::NotExist, otherwise returns back the enum in the Err case of the result

pub fn is_io(&self) -> bool

Returns true if this is a Error::Io, otherwise false

pub fn as_io_mut(&mut self) -> Option<&mut Error>

Optionally returns mutable references to the inner fields if this is a Error::Io, otherwise None

pub fn as_io(&self) -> Option<&Error>

Optionally returns references to the inner fields if this is a Error::Io, otherwise None

pub fn into_io(self) -> Result<Error, Self>

Returns the inner fields if this is a Error::Io, otherwise returns back the enum in the Err case of the result

pub fn is_host(&self) -> bool

Returns true if this is a Error::Host, otherwise false

pub fn as_host_mut(&mut self) -> Option<&mut HostError>

Optionally returns mutable references to the inner fields if this is a Error::Host, otherwise None

pub fn as_host(&self) -> Option<&HostError>

Optionally returns references to the inner fields if this is a Error::Host, otherwise None

pub fn into_host(self) -> Result<HostError, Self>

Returns the inner fields if this is a Error::Host, otherwise returns back the enum in the Err case of the result

pub fn is_other(&self) -> bool

Returns true if this is a Error::Other, otherwise false

pub fn as_other_mut(&mut self) -> Option<&mut Error>

Optionally returns mutable references to the inner fields if this is a Error::Other, otherwise None

pub fn as_other(&self) -> Option<&Error>

Optionally returns references to the inner fields if this is a Error::Other, otherwise None

pub fn into_other(self) -> Result<Error, Self>

Returns the inner fields if this is a Error::Other, otherwise returns back the enum in the Err case of the result