hyperactor_mesh::host

Function spawn_proc

pub async fn spawn_proc<A, S, F>(
    proc_id: ProcAddr,
    backend_addr: ChannelAddr,
    callback_addr: ChannelAddr,
    spawn: S,
) -> Result<Proc, HostError>
where
    A: Actor + Referable + Binds<A>,
    S: FnOnce(Proc) -> F,
    F: Future<Output = Result<ActorHandle<A>, Error>>,

Spawn a proc at proc_id with an A-typed agent actor, forwarding messages to the provided backend_addr, and returning the proc’s address and agent actor on the provided callback_addr.