hyperactor_mesh

Module logging

Structs

FileAppender

File appender, coordinates write access to a file via a channel.

FileMonitorMessage

Message sent to FileMonitor

LocalLogSender

Write the log to a local unix channel so some actors can listen to it and stream the log back.

LogClientActor

A client to receive logs from remote processes

LogForwardActor

A log forwarder that receives the log from its parent process and forward it back to the client

StreamFwder

Given a stream forwards data to the provided channel.

Enums

LogClientMessage

Messages that can be sent to the LogClient locally.

LogForwardMessage

Messages that can be sent to the LogForwarder

LogMessage

Messages that can be sent to the LogClientActor remotely.

OutputTarget

Represents the target output stream (stdout or stderr)

Stream

Statics

FORCE_FILE_LOG

If enabled, local logs are also written to a file and aggregated

PREFIX_WITH_RANK

Prefixes logs with rank

READ_LOG_BUFFER

Maximum number of lines to batch before flushing to client This means that stdout/err reader will be paused after reading HYPERACTOR_READ_LOG_BUFFER lines. After pause lines will be flushed and reading will resume.

Traits

LogClientMessageClient

The custom client trait for this message type.

LogClientMessageHandler

The custom handler trait for this message type.

LogForwardMessageClient

The custom client trait for this message type.

LogForwardMessageHandler

The custom handler trait for this message type.

LogMessageClient

The custom client trait for this message type.

LogMessageHandler

The custom handler trait for this message type.

LogSender

Trait for sending logs