hyperactor_mesh

Module proc_mesh

Structs

ProcMesh

A mesh of processes.

ProcMeshRef

A reference to a ProcMesh, consisting of a set of ranked ProcRefs, arranged into a region. ProcMeshes are named, uniquely identifying the ProcMesh from which the reference was derived.

ProcRef

A reference to a single hyperactor::Proc.

Constants

COMM_ACTOR_NAME

Name used for the mesh communication actor spawned on each user proc.

Statics

ACTOR_SPAWN_MAX_IDLE

The maximum idle time between updates while spawning actor meshes.

GET_ACTOR_STATE_MAX_IDLE

The maximum idle time between updates while waiting for a response to GetState from ProcAgent.