hyperactor_mesh::proc_mesh

Module mesh_agent

The mesh agent actor manages procs in ProcMeshes.

Structs

ActorSpec

Actor spec.

ActorState

Actor state.

GetProc

A handler to get a clone of the proc managed by this agent. This is used to obtain the local proc from a host mesh.

NewClientInstance

A local handler to get a new client instance on the proc. This is used to create root client instances.

ProcMeshAgent

A mesh agent is responsible for managing procs in a [ProcMesh].

Enums

GspawnResult

StopActorResult

Traits

GetProcClient

The custom client trait for this message type.

GetProcHandler

The custom handler trait for this message type.

MeshAgentMessageClient

The custom client trait for this message type.

MeshAgentMessageHandler

The custom handler trait for this message type.

NewClientInstanceClient

The custom client trait for this message type.

NewClientInstanceHandler

The custom handler trait for this message type.