hyperactor_mesh::proc_mesh

Struct ProcMesh

pub struct ProcMesh { /* private fields */ }

A ProcMesh maintains a mesh of procs whose lifecycles are managed by an allocator.

Implementations

impl ProcMesh

pub async fn allocate( alloc: impl Alloc + Send + Sync + 'static, ) -> Result<Self, AllocatorError>

pub fn allocate_boxed( alloc: Box<dyn Alloc + Send + Sync>, ) -> impl Future<Output = Result<Self, AllocatorError>>

Allocate a new ProcMesh from the provided allocator. Allocate returns after the mesh has been successfully (and fully) allocated, returning early on any allocation failure.

pub async fn spawn<A: RemoteSpawn, C: Actor>( &self, cx: &C, actor_name: &str, params: &A::Params, ) -> Result<RootActorMesh<'_, A>, Error>
where A::Params: RemoteMessage, C::A: Handler<MeshFailure>,

Spawn an ActorMesh by launching the same actor type on all agents, using the same parameters instance for every actor.

actor_name: Name for all spawned actors.

params: Reference to the parameter struct, reused for all actors.

Bounds:

A: Actor — we actually spawn this type on each agent.

A: Referable — we return a RootActorMesh<'_, A> that contains ActorRef<A>s; those exist only for A: Referable.

A::Params: RemoteMessage — params must be serializable to cross proc boundaries when launching each actor.

pub fn client(&self) -> &Instance<()>

A client actor used to communicate with any member of this mesh.

pub fn client_proc(&self) -> &Proc

pub fn proc_id(&self) -> &ProcId

pub fn world_id(&self) -> &WorldId

pub fn events(&mut self) -> Option<ProcEvents>

An event stream of proc events. Each ProcMesh can produce only one such stream, returning None after the first call.

pub fn shape(&self) -> &Shape

pub async fn stop_actor_by_name( &self, cx: &impl Actor, mesh_name: &str, reason: &str, ) -> Result<(), Error>

Send stop actors message to all mesh agents for a specific mesh name