hyperactor_mesh::resource

Module mesh

This module defines common types for mesh resources. Meshes are managed as resources, usually by a controller actor implementing the crate::resource behavior.

The mesh controller manages all aspects of the mesh lifecycle, and the owning actor uses the resource behavior directly to query the state of the mesh.

Structs

Spec

Mesh specs

State

Mesh states

Traits

Mesh

A mesh trait bundles a set of types that together define a mesh resource.