hyperactor_mesh::v1::actor_mesh

Struct ActorMeshRef

pub struct ActorMeshRef<A: Referable> { /* private fields */ }

A reference to a stable snapshot of an ActorMesh.

Implementations

impl<A: Referable> ActorMeshRef<A>

pub fn cast<M>(&self, cx: &impl Actor, message: M) -> Result<()>
where A: RemoteHandles<M> + RemoteHandles<IndexedErasedUnbound<M>>, M: Castable + RemoteMessage + Clone,

Cast a message to all the actors in this mesh

pub fn cast_for_tensor_engine_only_do_not_use<M>( &self, cx: &impl Actor, sel: Selection, message: M, ) -> Result<()>
where A: RemoteHandles<M> + RemoteHandles<IndexedErasedUnbound<M>>, M: Castable + RemoteMessage + Clone,

Cast a message to the actors in this mesh according to the provided selection. This should only be used for temporary support for selections in the tensor engine. If you use this for anything else, you will be fired (you too, OSS contributor).

pub async fn actor_states( &self, cx: &impl Actor, ) -> Result<ValueMesh<State<ActorState>>>

pub fn name(&self) -> &Name

pub fn proc_mesh(&self) -> &ProcMeshRef

pub fn controller(&self) -> &Option<ActorRef<ActorMeshController<A>>>

pub async fn next_supervision_event( &self, cx: &impl Actor, ) -> Result<MeshFailure, Error>

Returns the next supervision event occurring on this mesh. Await this simultaneously with the return result of a message (such as awaiting a reply after a cast) to get back a message that indicates the actor that failed, instead of waiting forever for a reply. If there are multiple simultaneous awaits of next_supervision_event, all of them will receive the same event.

pub fn clone_with_supervision_receiver(&self) -> Self

Same as Clone, but includes a shared supervision receiver. This copy will share the same health state and get the same supervision events. Will have a separate cache.