hyperactor_mesh::v1

Module host_mesh

Re-exports

pub use crate::v1::host_mesh::mesh_agent::HostMeshAgent;

Modules

mesh_agent

The mesh agent actor that manages a host.

Structs

HostMesh

An owned mesh of hosts.

HostMeshRef

A non-owning reference to a mesh of hosts.

HostRef

A reference to a single host.

Enums

HostMeshRefParseError

The type of error occuring during HostMeshRef parsing.

Statics

GET_PROC_STATE_MAX_IDLE

The maximum idle time between updates while querying host meshes for their proc states.

PROC_SPAWN_MAX_IDLE

The maximum idle time between updates while spawning proc meshes.

PROC_STOP_MAX_IDLE

The maximum idle time between updates while stopping proc meshes.