hyperactor_mesh::v1::host_mesh

Module mesh_agent

The mesh agent actor that manages a host.

Structs

GetHostMeshAgent

GetLocalProc

A local-only message to access the “local” proc on the host. This is used to bootstrap the root mesh process client on the local singleton host mesh.

HostMeshAgent

A mesh agent is responsible for managing a host iny a [HostMesh], through the resource behaviors defined in crate::resource.

ProcState

ShutdownHost

Enums

HostAgentMode

Represents the different ways a Host can be managed by an agent.

Traits

GetHostMeshAgentClient

The custom client trait for this message type.

GetHostMeshAgentHandler

The custom handler trait for this message type.

GetLocalProcClient

The custom client trait for this message type.

GetLocalProcHandler

The custom handler trait for this message type.

ShutdownHostClient

The custom client trait for this message type.

ShutdownHostHandler

The custom handler trait for this message type.