hyperactor_mesh::v1

Module mesh_controller

Structs

ActorMeshController

The implementation of monitoring works as follows:

CheckState

Check state of the actors in the mesh. This is used as a self message to periodically check.

Subscribe

Subscribe me to updates about a mesh. If a duplicate is subscribed, only a single message is sent. Will send None if there are no failures on the mesh periodically. This guarantees the listener that the controller is still alive. Make sure to filter such events out as not useful.

Unsubscribe

Unsubscribe me to future updates about a mesh. Should be the same port used in the Subscribe message.

Statics

ACTOR_MESH_SUBSCRIBER_MESSAGE

If present in a message header, the message is from an ActorMeshController to a subscriber and can be safely dropped if it is returned as undeliverable.

SUPERVISION_POLL_FREQUENCY

Time between checks of actor states to create supervision events for owners. The longer this is, the longer it will take to detect a failure and report it to all subscribers; however, shorter intervals will send more frequent messages and heartbeats just to see everything is still running. The default is chosen to balance these two objectives. This also controls how frequently the healthy heartbeat is sent out to subscribers if there are no failures encountered.