Crate hyperactor_mesh

This crate provides hyperactor’s mesh abstractions.

Re-exports

pub use actor_mesh::RootActorMesh;

pub use actor_mesh::SlicedActorMesh;

pub use bootstrap::Bootstrap;

pub use bootstrap::bootstrap;

pub use bootstrap::bootstrap_or_die;

pub use comm::CommActor;

pub use mesh::Mesh;

pub use proc_mesh::ProcMesh;

pub use proc_mesh::SlicedProcMesh;

pub use dashmap;

Modules

actor_mesh

alloc

This module defines a proc allocator interface as well as a multi-process (local) allocator, ProcessAllocator.

bootstrap

comm

config

Configuration for Hyperactor Mesh.

connect

Actor-based duplex bytestream connections.

logging

mesh

mesh_selection

proc_mesh

reference

resource

This modules defines a set of common message types used for managing resources in hyperactor meshes.

router

This module supports (in-process) global routing in hyperactor meshes.

selection

Selection algebra for describing multidimensional mesh regions. This module defines a recursive algebra for selecting coordinates in a multidimensional space.

shape

Core types for representing multidimensional shapes and strides.

shared_cell

shortuuid

This module implements a “short” (64-bit) UUID, used to assign names to various objects in meshes.

supervision

Messages used in supervision of actor meshes.

test_utils

v1

A temporary holding space for APIv1 of the Hyperactor Mesh. This will be moved down to the base module when we graduate the APIs and fully deprecate the “v0” APIs.

Macros

extent

Construct a new extent with the given set of dimension-size pairs.

mesh_id

ok

sel

Parse a compact selection expression into a [Selection]. See [selection::parse] for syntax documentation.

sel_from_shape

Construct a Selection from a Shape using labeled dimension constraints.

shape

Construct a new shape with the given set of dimension-size pairs in row-major order.