hyperactor_mesh

Module actor_mesh

Structs

ActorSupervisionEvents

Supervision event stream for actor mesh. It emits actor supervision events.

RootActorMesh

A mesh of actor instances. ActorMeshes are obtained by spawning an actor on a ProcMesh.

SlicedActorMesh

Enums

CastError

The type of error of casting operations.

Statics

CAST_ACTOR_MESH_ID

Which mesh this message was cast to. Used for undeliverable message handling, where the CastMessageEnvelope is serialized, and its content cannot be inspected.

Traits

ActorMesh

A mesh of actors, all of which reside on the same ProcMesh.

Functions

update_undeliverable_envelope_for_casting

An undeliverable might have its sender address set as the comm actor instead of the original sender. Update it based on the headers present in the message so it matches the sender.