hyperactor_mesh::bootstrap

Function host

pub async fn host(
    addr: ChannelAddr,
    command: Option<BootstrapCommand>,
    config: Option<Attrs>,
    exit_on_shutdown: bool,
) -> Result<(ActorHandle<HostAgent>, HostShutdownHandle)>

Bootstrap a host in this process, returning a handle to the mesh agent.

To obtain the local proc, use GetLocalProc on the returned host mesh agent, then use GetProc on the returned proc mesh agent.

addr: the listening address of the host; this is used to bind the frontend address;

command: optional bootstrap command to spawn procs, otherwise [BootstrapProcManager::current];

config: optional runtime config overlay.

exit_on_shutdown: if true, HostShutdownHandle::join will call process::exit after draining.