hyperactor_mesh

Module proc_mesh

Modules

mesh_agent

The mesh agent actor manages procs in ProcMeshes.

Structs

GlobalClientActor

ProcEvents

An event stream of ProcEvent

ProcMesh

A ProcMesh maintains a mesh of procs whose lifecycles are managed by an allocator.

SlicedProcMesh

Enums

ProcEvent

Proc lifecycle events.

SupervisionEventState

Statics

DEFAULT_TRANSPORT

Default transport type to use across the application.

Traits

SharedSpawnable

Spawns from shared (Arc) proc meshes, providing [ActorMesh]es with static lifetimes.

Functions

default_bind_spec

Get the default bind spec to use across the application.

default_transport

Temporary: used to support the legacy allocator-based V1 bootstrap. Should be removed once we fully migrate to simple bootstrap.

global_root_client

Context use by root client to send messages. This mailbox allows us to open ports before we know which proc the messages will be sent to.